Refactor the `arti` primitive codebase
