+++ gi_loadouts/type/arti/base.py
