+++ gi_loadouts/type/arti/stat.py
